requirements.txt
snowflake-snowpark-python[pandas]
snowflake-ml-python
pyyaml
pandas
scikit-learn
pytest